+++ src/game/Result.js
